[DEVMAJOR] Allow pomp_loop_get_fd to work for win32
- The return type is intptr_t to return a HANDLE to an event
- The first time this is called, an internal thread is created
  that do the wait on all internal handles and notify a single
  event when ready.
- Then pomp_loop_process_fd (or pomp_loop_wait_and_process with 0 timeout)
  can be called as usual.
- For that to work, all waitable handles should not reset themselves
  automatically. That's why the timer implementation goes back to
  timer queue instead of waitable timers.
- The same kind of implementation could be done for poll to simulate
  epoll but this is not done here.
- The returned handle can be integrated in a qt main loop with a
  QWinEventNotifier

+++ src/pomp_ctx.c
@@ -951,7 +951,7 @@ int pomp_ctx_stop(struct pomp_ctx *ctx)
 /*
  * See documentation in public header.
  */
-int pomp_ctx_get_fd(const struct pomp_ctx *ctx)
+intptr_t pomp_ctx_get_fd(const struct pomp_ctx *ctx)
 {
 	POMP_RETURN_ERR_IF_FAILED(ctx != NULL, -EINVAL);
 	return pomp_loop_get_fd(ctx->loop);
