+++ src/pomp_loop_linux.c
@@ -230,7 +230,7 @@ static int pomp_loop_epoll_do_remove(struct pomp_loop *loop,
 /**
  * @see pomp_loop_do_get_fd.
  */
-static int pomp_loop_epoll_do_get_fd(struct pomp_loop *loop)
+static intptr_t pomp_loop_epoll_do_get_fd(struct pomp_loop *loop)
 {
 	return loop->efd;
 }
